requirements.txt
# Core dependencies
pyyaml>=6.0.1
pydantic>=2.5.0
xmltodict>=0.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# dbt
dbt-core>=1.7.0
dbt-duckdb>=1.7.0

# Data validation
great-expectations>=0.18.0

# Utilities
python-dateutil>=2.8.2
